Fix #209: add selectable jupyterhub authenticator (#211)
There was a bug in an update to the sirepo jupyterhub authenticator
 (original change: git.radiasoft.org/sirepo/pull/4211 subsequent bug: git.radiasoft.org/sirepo/issues/4583).
In order to roll back the change we wanted to try making jupyterhublogin a default_proprietary_sim_type
again. But, that forced the use of the old authenticator class (`sirepo.jupyterhub.Authenticator`) which
doesn't exist in the latest image. So, add this configuration so we can say which authenticator
we want regardless of which sim type jupyterhublogin is.