Fix #373: allow specification of docker_image_is_local for devbox (#374)
Use component module_name (devbox) instead of name (devbox_{user_name}) when trying
to figure out docker_image_is_local.

This assumes all devbox's on a host are either local or not. For now, this assumption
works but it may be we need per user docker_image_is_local configuration in the future.